chore(array-utils): move plugins to a separate file
